requirements.txt
aiofiles==0.8.0
boto3==1.20.35
moto==2.3.1
pytest==6.2.5
mysql-connector-python==8.0.27
